+++ src/analytics/PivotTable.js
@@ -5,7 +5,7 @@ import 'react-pivottable/pivottable.css';
 
 const PivotTable = ({ cards }) => {
   const data = cards.map((card) => ({
-    CMC: card.cmc || card.details.cmc,
+    CMC: card.cmc ?? card.details.cmc,
     Color: card.colors.join(),
     Finish: card.finish,
     'Type Line': card.type_line,
